+++ tests/E2E/test/campaigns/high/08_stocks/1_group_of_product_quantity.js
@@ -8,11 +8,13 @@ let promise = Promise.resolve();
 
 productData = [{
   name: 'FirstProduct',
+  reference: 'firstProduct',
   quantity: "100",
   price: '5',
   image_name: 'image_test.jpg',
 }, {
   name: 'SecondProduct',
+  reference: 'secondProduct',
   quantity: "100",
   price: '5',
   image_name: 'image_test.jpg',
