+++ tests/E2E/test/campaigns/high/15_installation/1_installation_language_equal_to_country.js
