+++ tests/E2E/test/campaigns/high/15_installation/2_installation_language_different_to_country.js
