+++ tests/E2E/test/campaigns/high/15_installation/common_installation.js
