+++ tests/E2E/test/campaigns/regular/01_install.js
@@ -1,5 +1,5 @@
 require('../../globals.webdriverio.js');
 
 if (typeof install_shop !== 'undefined' && install_shop) {
-  require('../high/14_installation/2_installation_language_different_to_country.js');
+  require('../high/15_installation/2_installation_language_different_to_country.js');
 }
